clean-up and fix SkeletonRendererMaterialOverride
