--- spine-unity/Assets/spine-unity/Modules/MaterialOverride/Editor/SkeletonRendererMaterialOverrideInspector.cs
@@ -1,19 +1,0 @@
-﻿using UnityEditor;
-
-[CustomEditor(typeof(SkeletonRendererMaterialOverride))]
-public class SkeletonRendererMaterialOverrideInspector : Editor {
-	public override void OnInspectorGUI() {
-		SkeletonRendererMaterialOverride obj = (SkeletonRendererMaterialOverride) target;
-
-		EditorGUI.BeginChangeCheck();
-		{
-			base.OnInspectorGUI();
-		}
-		if (EditorGUI.EndChangeCheck()) {
-			obj.RemoveAtlasMaterialOverrides();
-			obj.RemoveSlotMaterialOverrides();
-			obj.SetAtlasMaterialOverrides();
-			obj.SetSlotMaterialOverrides();
-		}
-	}
-}
